TODO tuning
Regarding the test, I tested it and it works well enough. I think the
imperfections are related to VSCode rather than to this extension so
let's just ignore them for now. Fixing them would be a separate project
inside of the VSCode project. Maybe we should report it somehow?

Adding that Quick Fix for JIRA issue ID styles would be fun, but nobody
would ever use it so never mind.