Remove problematic ruler
In the VSCode repo, a ruler was added at column 50:
https://github.com/microsoft/vscode/pull/201091

That ruler is causing problems with the rewrap extension, and it's
unclear to me what the benefit of it is that outweighs the drawbacks.

Please let me know if you know how that ruler is useful!

I'd love to hear about it, and depending on the use case I might
reconsider this override.

Workaround for: https://github.com/stkb/Rewrap/issues/395